requirements.txt
backoff==1.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
docutils==0.18.1
idna==3.3
opencage==2.0.0
py4j==0.10.9.3
pycparser==2.21
pygeohash==1.2.0
pyOpenSSL==22.0.0
pyspark==3.2.1
python-geohash==0.8.5
requests==2.27.1
urllib3==1.26.9
venv-pack==0.2.0
